pyproject.toml
[project]
name = "app-vacancy-analysis"
version = "0.1.0"
description = ""
authors = [
    { name = "MrchknRV", email = "rodionmarochkin32@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
flake8 = "^7.3.0"
mypy = "^1.17.1"
black = "^25.1.0"
isort = "^6.0.1"


[tool.poetry.group.dev.dependencies]
requests = "^2.32.4"
psycopg2-binary = "^2.9.10"


[tool.poetry.group.test.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
python-dotenv = "^1.1.1"

[tool.black]
line-length = 119
exclude = "venv"

[tool.isort]
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = "venv"